Move all non game related modules out of game folder
update header references
