+++ xlive/Blam/Engine/hs/hs.h
