+++ xlive/Blam/Engine/objects/damage.h
