+++ xlive/H2MOD/Modules/CustomVariantSettings/CustomVariantSettings.cpp
@@ -4,7 +4,7 @@
 
 #include "Blam\Engine\memory\bitstream.h"
 #include "Blam/Engine/game/GameTimeGlobals.h"
-#include "Blam/Engine/game/PhysicsConstants.h"
+#include "Blam/Engine/physics/physics_constants.h"
 #include "Blam\Engine\Players\PlayerControls.h"
 #include "Blam\Engine\Networking\Session\NetworkSession.h"
 #include "Blam\Engine\Networking\NetworkMessageTypeCollection.h"
